[refactor] moved side bar logic into seperate class
